KA-161 Rewrite reducer to correspond with nesting
