+++ onboarding-task/src/reducers/items/items.js
@@ -1,27 +1,11 @@
-import { OrderedMap } from 'immutable';
-import {
-  CREATE_ITEM,
-  DELETE_ITEM,
-  UPDATE_ITEM,
-} from '../../constants/actionTypes';
-import { item } from './item';
+import { combineReducers } from 'redux';
 
-export const items = (state = new OrderedMap(), action) => {
-  switch (action.type) {
-    case CREATE_ITEM: {
-      const newItem = action.payload.item;
+import { data } from './data/data';
+import { flags } from './flags/flags';
+import { ids } from './ids';
 
-      return state.set(newItem.id, newItem);
-    }
-    case UPDATE_ITEM: {
-      const existingItem = state.get(action.payload.item.id);
-      const newItem = item(existingItem, action);
-
-      return state.set(action.payload.item.id, newItem);
-    }
-    case DELETE_ITEM:
-      return state.delete(action.payload.itemId);
-    default:
-      return state;
-  }
-};
+export const items = combineReducers({
+  ids,
+  data,
+  flags,
+});
