+++ onboarding-task/src/reducers/reducer.js
@@ -1,13 +1,6 @@
 import { combineReducers } from 'redux';
-import { items as itemReducer } from './items/items';
-import { flags } from './flags/flags';
-import { ids } from './ids';
 
-const items = combineReducers({
-  ids,
-  byIds: itemReducer,
-  flags,
-});
+import { items } from './items/items';
 
 export const rootReducer = combineReducers({
   items,
