+++ onboarding-task/test/reducers/updateFlagsReducer.test.js
@@ -1,5 +1,5 @@
 import { ListItemFlag } from '../../src/models/ListItemFlag';
-import { flag } from '../../src/reducers/flags/flag';
+import { flag } from '../../src/reducers/items/flags/flag';
 import { toggleBeingEdited } from '../../src/actions/userActions';
 
 describe('Update Flags Reducer', () => {
